chore: moves all entrypoints to a new directory
Moves autosync entrypoint to entrypoints directory
Removes infra directory
