+++ tests/trestlebot/__init__.py
@@ -13,4 +13,4 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
-"""Test package. for trestlebot top-level logic."""
+"""Test package for trestlebot top-level logic."""
